+++ packages/calcite-components/src/components/date-picker-month/date-picker-month.tsx
@@ -540,17 +540,16 @@ export class DatePickerMonth {
           range={!!this.startDate && !!this.endDate && !sameDate(this.startDate, this.endDate)}
           rangeEdge={dayInWeek === 0 ? "start" : dayInWeek === 6 ? "end" : undefined}
           rangeHover={this.isRangeHover(date)}
-          scale={this.scale}
-          selected={this.isSelected(date)}
-          startOfRange={this.isStartOfRange(date)}
-          value={date}
-          // eslint-disable-next-line react/jsx-sort-props -- ref should be last so node attrs/props are in sync (see https://github.com/Esri/calcite-design-system/pull/6530)
           ref={(el: HTMLCalciteDatePickerDayElement) => {
             // when moving via keyboard, focus must be updated on active date
             if (ref && active && this.activeFocus) {
               el?.setFocus();
             }
           }}
+          scale={this.scale}
+          selected={this.isSelected(date)}
+          startOfRange={this.isStartOfRange(date)}
+          value={date}
         />
       </div>
     );