+++ packages/calcite-components/src/components/list/list.tsx
@@ -47,7 +47,13 @@ import {
   T9nComponent,
   updateMessages,
 } from "../../utils/t9n";
-import { NumberingSystem, numberStringFormatter } from "../../utils/locale";
+import {
+  connectLocalized,
+  disconnectLocalized,
+  LocalizedComponent,
+  NumberingSystem,
+  numberStringFormatter,
+} from "../../utils/locale";
 import { CSS, debounceTimeout, SelectionAppearance, SLOTS } from "./resources";
 import { ListMessages } from "./assets/list/t9n";
 import { ListDragDetail } from "./interfaces";
@@ -71,7 +77,12 @@ const parentSelector = "calcite-list-item-group, calcite-list-item";
   assetsDirs: ["assets"],
 })
 export class List
-  implements InteractiveComponent, LoadableComponent, SortableComponent, T9nComponent
+  implements
+    InteractiveComponent,
+    LoadableComponent,
+    LocalizedComponent,
+    SortableComponent,
+    T9nComponent
 {
   // --------------------------------------------------------------------------
   //
@@ -254,12 +265,13 @@ export class List
    */
   @Event({ cancelable: false }) calciteInternalListDefaultSlotChange: EventEmitter<void>;
 
-  @Listen("calciteListItemToggle")
-  handleCalciteListItemToggle(): void {
+  @Listen("calciteInternalListItemToggle")
+  handleCalciteListItemToggle(event: CustomEvent): void {
     if (this.parentListEl) {
       return;
     }
 
+    event.stopPropagation();
     this.borderItems();
   }
 
@@ -283,7 +295,7 @@ export class List
 
   @Listen("calciteInternalListItemActive")
   handleCalciteInternalListItemActive(event: CustomEvent): void {
-    if (!!this.parentListEl) {
+    if (this.parentListEl) {
       return;
     }
 
@@ -298,7 +310,7 @@ export class List
 
   @Listen("calciteListItemSelect")
   handleCalciteListItemSelect(): void {
-    if (!!this.parentListEl) {
+    if (this.parentListEl) {
       return;
     }
 
@@ -313,7 +325,7 @@ export class List
 
   @Listen("calciteHandleNudge")
   handleCalciteHandleNudge(event: CustomEvent<HandleNudge>): void {
-    if (!!this.parentListEl) {
+    if (this.parentListEl) {
       return;
     }
 
@@ -322,7 +334,7 @@ export class List
 
   @Listen("calciteInternalListItemSelect")
   handleCalciteInternalListItemSelect(event: CustomEvent): void {
-    if (!!this.parentListEl) {
+    if (this.parentListEl) {
       return;
     }
 
@@ -343,7 +355,7 @@ export class List
       selectMultiple: boolean;
     }>,
   ): void {
-    if (!!this.parentListEl) {
+    if (this.parentListEl) {
       return;
     }
 
@@ -368,7 +380,7 @@ export class List
 
   @Listen("calciteInternalListItemChange")
   handleCalciteInternalListItemChange(event: CustomEvent): void {
-    if (!!this.parentListEl) {
+    if (this.parentListEl) {
       return;
     }
 
@@ -392,6 +404,7 @@ export class List
       return;
     }
 
+    connectLocalized(this);
     connectMessages(this);
     this.connectObserver();
     this.updateListItems();
@@ -421,6 +434,7 @@ export class List
     this.disconnectObserver();
     disconnectSortableComponent(this);
     disconnectInteractive(this);
+    disconnectLocalized(this);
     disconnectMessages(this);
   }
 
@@ -546,9 +560,8 @@ export class List
                         items={dataForFilter}
                         onCalciteFilterChange={this.handleFilterChange}
                         placeholder={filterPlaceholder}
-                        value={filterText}
-                        // eslint-disable-next-line react/jsx-sort-props -- ref should be last so node attrs/props are in sync (see https://github.com/Esri/calcite-design-system/pull/6530)
                         ref={this.setFilterEl}
+                        value={filterText}
                       />
                       <slot
                         name={SLOTS.filterActionsEnd}
@@ -846,7 +859,7 @@ export class List
       item.dragHandle = dragEnabled;
     });
 
-    if (!!this.parentListEl) {
+    if (this.parentListEl) {
       this.setUpSorting();
       return;
     }